memdag -> perdag move delete client head etc to replicache
